+++ LunaDll/SMBXInternal/Sound.h
@@ -3,10 +3,10 @@
 #include <string>
 
 namespace SMBXSound {
-	void PlaySFX(int soundindex);
+	void PlaySFX(short soundindex);
 
 	void SetMusic(int musicnum, int section);
 	void SetMusicPath(int section, std::wstring path);
-	void PlayMusic(int section);
+	void PlayMusic(short section);
 }
 #endif
